+++ h2/src/main/org/h2/command/dml/ExecuteProcedure.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * Copyright 2004-2021 H2 Group. Multiple-Licensed under the MPL 2.0,
  * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
@@ -68,7 +68,7 @@ public class ExecuteProcedure extends Prepared {
     }
 
     @Override
-    public ResultInterface query(int limit) {
+    public ResultInterface query(long limit) {
         setParameters();
         Prepared prepared = procedure.getPrepared();
         return prepared.query(limit);
