+++ h2/src/main/org/h2/expression/condition/UniquePredicate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2020 H2 Group. Multiple-Licensed under the MPL 2.0,
+ * Copyright 2004-2021 H2 Group. Multiple-Licensed under the MPL 2.0,
  * and the EPL 1.0 (https://h2database.com/html/license.html).
  * Initial Developer: H2 Group
  */
@@ -41,9 +41,9 @@ public class UniquePredicate extends PredicateWithSubquery {
         }
 
         @Override
-        public int getRowCount() {
+        public long getRowCount() {
             // Not required
-            return 0;
+            return 0L;
         }
 
         @Override
@@ -59,7 +59,7 @@ public class UniquePredicate extends PredicateWithSubquery {
             if (values.length != columnCount) {
                 values = Arrays.copyOf(values, columnCount);
             }
-            int expected = result.getRowCount() + 1;
+            long expected = result.getRowCount() + 1;
             result.addRow(values);
             if (expected != result.getRowCount()) {
                 hasDuplicates = true;
